Update citizend sale page
